requirements.txt
pyspark==3.3.2
apache-airflow-providers-apache-spark

